Update convertFormat as a cypher function
